fix: ensure ECL support
Two things to note here:

The `equal` behaviour for pathnames is different between SBCL and ECL.
The latter seems to take `:version` into account, while the former does
not.

`make-pathname` under SBCL seems to automatically filter out any empty
strings, but ECL does not, so we have to do that manually.